Update ptBR.lua
In line 33, i did EN normal some times in my main and i think this is the best way to describe this for brazilians.
Just to note, i entered LFR to test this function, i was promoted, but my markers were conflicting with someone's DBM skull marker. No one cared because it was LFR, they ignore the fight, but imo BigWigs double markers usage is way better when the mechanic is really needed.