added option to reverse the order bars are stacked
Bars usually grow away from anchor depending on time left. With this change there is the option to have bars running out soon close to the anchor and those that have a lot of time left to be most distant to anchor. Some configurations need this for example when you want the bar-stack to grow upwards but have the shortest bar on top for better visibility in screen region.