Update koKR.Lua
In Addition to this, I would like to propose the 
--L.room = "room (%d/8)"
to change to something like 
--L.room = "room number %d"
so in korean, it would be
--L.room = "%d 번 방 
Because it would be the more common way of saying it.
maybe for the 7th room, you could have some text to tell players to pick up gems as well? 

I left some ambiguous parts blank, such as "hand: %s " and "bulwark: %s". After this week's raid I'll try and update it.