Updates Wrath timers
Wraths occur with a constant but aspect specific offset to the spawn timers. Initially, spawn timers are 30 seconds. Spawn timers are reduced to 20 and 15 seconds after one and two bosses die, respectively. From the active aspects, the one least recently seen spawns next. Spawn timers continue when a boss dies, hence the time between the same Wraths can be extended from 30 + 30 = 20 + 20 + 20 = 15 + 15 + 15 + 15 = 60 to 30 + 20 +20 = 70 or 20 + 15 + 15 + 15 =65 seconds.
With this change, Wrath timers are shorter and only one timer is active at a time. However, the minimal duration for Paku's Wrath is 13 + 15 = 28 seconds, which is long enough to prepare.